Address some review feedback
- Avoids some of the inverted logic in ifdefs
- Conditionally run the zipkin python integration test
- Adds a TODO for adding static module registration
