+++ test/integration/BUILD
@@ -8,14 +8,13 @@ load("@python_pip_deps//:requirements.bzl", "requirement")
 
 envoy_package()
 
-py_test(
-    name = "integration_test",
+py_library(
+    name = "integration_test_base",
     srcs = [
         "common.py",
-        "integration_test.py",
         "integration_test_fixtures.py",
         "nighthawk_test_server.py",
-        "test_integration_basics.py",
+        "utility.py",
     ],
     data = [
         "configurations/nighthawk_http_origin.yaml",
@@ -24,8 +23,6 @@ py_test(
         "//:nighthawk_test_server",
         "@envoy//test/config/integration/certs",
     ],
-    python_version = "PY3",
-    srcs_version = "PY3ONLY",
     deps = [
         requirement("requests"),
         # The following are implied by 'request'.
@@ -49,5 +46,34 @@ py_test(
         requirement("pytest-xdist"),
         requirement("execnet"),
         requirement("apipkg"),
+    ],
+)
+
+py_library(
+    name = "test_integration_basics_lib",
+    srcs = ["test_integration_basics.py"],
+    deps = [":integration_test_base"],
+)
+
+py_library(
+    name = "test_integration_zipkin_lib",
+    srcs = ["test_integration_zipkin.py"],
+    deps = [":integration_test_base"],
+)
+
+py_test(
+    name = "integration_test",
+    srcs = ["integration_test.py"],
+    args = select({
+        "//bazel:zipkin_disabled": [
+            "\"not zipkin\"",
+        ],
+        "//conditions:default": [],
+    }),
+    python_version = "PY3",
+    srcs_version = "PY3ONLY",
+    deps = [
+        ":test_integration_basics_lib",
+        ":test_integration_zipkin_lib",
     ],
 )
