+++ test/integration/integration_test.py
@@ -9,5 +9,7 @@ import pytest
 
 if __name__ == '__main__':
   path = os.path.dirname(os.path.realpath(__file__))
-  r = pytest.main(["--rootdir=" + path, "-x", path, "-n", "20"], plugins=["xdist"])
+  test_selection_arg = sys.argv[1] if len(sys.argv) > 1 else ""
+  r = pytest.main(["--rootdir=" + path, "-k", test_selection_arg, "-x", path, "-n", "20"],
+                  plugins=["xdist"])
   exit(r)
