+++ test/integration/utility.py
@@ -1,0 +1,22 @@
+def assertEqual(a, b):
+  assert a == b
+
+
+def assertGreater(a, b):
+  assert a > b
+
+
+def assertGreaterEqual(a, b):
+  assert a >= b
+
+
+def assertLessEqual(a, b):
+  assert a <= b
+
+
+def assertNotIn(a, b):
+  assert a not in b
+
+
+def assertIn(a, b):
+  assert a in b
